+++ src/main/java/com/manhduc205/ezgear/dtos/PurchaseOrderDTO.java
@@ -16,8 +16,8 @@ public class PurchaseOrderDTO {
     private String supplierName;
     private Long warehouseId;
     private String status;
-    private BigDecimal subtotal;
-    private BigDecimal total;
+    private Long  subtotal;
+    private Long  total;
     private String note;
     private Long createdBy;
     private List<PurchaseOrderItemDTO> items;
